Generate Documentation; Reorganize Cobra Commands
Generate documentation from cobra commands.  Put the cobra commands into
a separate package so it can be imported for document generation.
