+++ docs/batchforce.md
@@ -1,0 +1,63 @@
+## batchforce
+
+Use Bulk API to update Salesforce records
+
+### Synopsis
+
+
+	Insert/Update/Delete Salesforce records using the Bulk API and a SOQL query.
+
+	Optionally use anonymous apex to provide additional context.
+
+	The SOQL query is used to generate the input.  Each record returned by the
+	query is made available to the Expr expression as a map named "record".  See
+	https://expr.medv.io/ for details on the Expr language.  The expression should
+	evaluate to an map of the form, "{ Field: Value, ... }" or an array of such
+	maps.
+
+	In addition to Expr's built-in operators and functions, the following
+	functions can be used within the expression:
+	- stripHtml: removes HTML tags
+	- escapeHtml: escapes characters using HTML entities like Apex's
+	  String.escapeHtml4 method
+	- escapeUnicode: escapes characters using Unicode escape sequences
+	  like Apex's String.escapeUnicode method
+	- base64: base-64 encodes input
+	- compareAndSet: check if key maps to value; if key doesn't exist, set it to
+	  value
+	- incr: increments the number stored at key by one. set to 0 if not set.
+
+	The + and - operators can be used to add, update, or remove fields on the
+	record object.  For example:
+	record + {RecordTypeId: apex.myRecordTypeId} - "RecordType.Name"
+
+	Additional context to be provided to the Expr expression by passing the
+	--context parameter containining anonymous apex to execute before the
+	records are queried.  Each apex variable defined will be available within
+	the "apex" map.
+
+	A csv file can be used as input instead of a SOQL query by using the --file
+	parameter.  This is often useful when combined with --apex to map input to
+	org-specific values such as Record Type Ids.
+	
+
+```
+batchforce
+```
+
+### Options
+
+```
+  -a, --account username   account username to use
+  -h, --help               help for batchforce
+      --quiet              supress informational log messages
+```
+
+### SEE ALSO
+
+* [batchforce delete](batchforce_delete.md)	 - delete Salesforce records using the Bulk API
+* [batchforce insert](batchforce_insert.md)	 - insert Salesforce records using the Bulk API
+* [batchforce update](batchforce_update.md)	 - Update Salesforce records using the Bulk API
+* [batchforce upsert](batchforce_upsert.md)	 - upsert Salesforce records using the Bulk API
+* [batchforce version](batchforce_version.md)	 - Display current version
+
