+++ docs/batchforce_version.md
@@ -1,0 +1,33 @@
+## batchforce version
+
+Display current version
+
+```
+batchforce version [flags]
+```
+
+### Examples
+
+```
+
+  batchforce version
+
+```
+
+### Options
+
+```
+  -h, --help   help for version
+```
+
+### Options inherited from parent commands
+
+```
+  -a, --account username   account username to use
+      --quiet              supress informational log messages
+```
+
+### SEE ALSO
+
+* [batchforce](batchforce.md)	 - Use Bulk API to update Salesforce records
+
